Move code related to step mode to dedicated module
